Refactor News_section component file structure and import paths
