+++ sents-web/src/components/admin/companyCard/index.tsx
